Grant view permissions for all CRTBs
Problem:
Not all roleTemplates include the provisioning cluster so when a CRTB is
created it's not guaranteed that the subject can see the provisioning
cluster

Solution:
If any CRTB is created, the subject for that CRTB needs to see the
provisioning cluster so grant view permsissions when the roleTemplate
does not normally grant it. This is how permissions work for management
clusters as well.