Get clusterName from rke-machine-config
Problem:
rke-machine-config does not return a clustername becuase it does not
live in the expected locations on the object

Solution:
Update getObjectClusterNames to look for the group of the object and
then find the owner reference for the provisioning cluster. These
objects are dynamic so the switch won't work as there is no concrete
type and they don't have a spec so it can't be added there on creation.