Retrieve the UserContext for cluster dynamically
When the cluster context is set for the proxy server at initialization,
it doesn't always refresh when a new request comes in. This can cause
the cache to become stale and the proxy server handler can fail to
fetch service accounts from it even though they exist in the downstream
cluster. This change pushes the cluster context getter down to the
request handling layer instead of the server creation layer so that it
is refreshed on every request.

(cherry picked from commit f1d453935c9d0c2dd3b3630687acebcdef4d412c)