Set InfrastructureReady condition on machine deletion
A previous change mistakenly set the Provision status when a machine was
being deleted. This change now updates the InfrastructureReady condition
when it has been detected that the infrastructure object has been
deleted.

This is necessary because the CAPI machine controller is waiting for the
infrastructure object to be deleted and this is required so that the
CAPI controller runs again.