Fix provisioning machine files permissions
A previous update changed the machine-provisioning job to run as a
non-root user, as well as updating permissions to volumes to ensure the
new user can access the appropriate files. There was one volume that was
missed, and this change updates the permissions of that volume to allow
the machine-provisioning job to access the files it needs.