Add RBAC rule for cleanup job user
The cleanup job service account now needs to be able to look up the
presence of the rancher service in the cluster and fails to perform any
cleanup if it does not have permission. This change adds the appropriate
RBAC rule to the main cluster role for the cleanup job.

(cherry picked from commit ae94d4940ffbbe77e56d2a2e17806c02bf825c2b)