Delete provisioning cluster after CAPI objects
The code seemed to intend to wait for the CAPI objects to be cleaned up
before deleting the provisioning cluster. However, the provisioning
cluster was deleted immediately. This caused issues when creating a
cluster with the same name after deletion.

In addition, if the provisioning cluster is in a bad state on creation,
then the CAPI objects will get created, but the finalizer will not be
put on the provisioning cluster. Therefore, it will be impossible to
delete the CAPI objects on provisioning cluster deletion. Instead, the
controller will ensure that the provisioning cluster can be successfully
updated in its current state before trying to apply the CAPI objects.

Lastly, if the machinestatus controller detects that the bootstrap
object has been deleted, then it will update the status of the machine
to reflect that and allow the CAPI machine controller to continue with
deletion.