Be more particular when setting node-external-ip
A previous change set the node-external-ip if the user did not provide
one. Instead, the node-external-ip is only set if a cloud provider is
not provided and if an IP address and private IP address can be
identified and are different. In addition, if the node-external-ip
address is set, then the advertise-address is set to the internal
address to maintain consistency.