Fix worker config and refactor IP address setting
First, and foremost, this change is to address the worker config getting
an invalid setting: advertise-address is a server only setting.

Secondly, this refactors the code that sets the IP addresses so that
advertise address and tls-san are set whether or not a cloud-provider has been
chosen. This is to ensure that the internal address is always used for
the apiserver IP.