Distinguish imported RKE2 clusters from provisioning clusters
A previous attempt was made to determine when a v3.Cluster upgrade
should be administrated by management or provisioning controllers. That
change was not able to distinguish imported RKE2 clusters where should be
administrated by the management controllers.

This change adds an annotation to the v3.Cluster object to better
distinguish these scenarios.