Fix imported cluster cleanup race
There is a race condition where the cluster object on the local cluster
is being deleted, but the cluster agent on the downstream cluster is not
yet cleaned up, and without this change the controller won't get a
kubeconfig for the deleted cluster and won't be able to create the
cleanup job. This change expands the exception made for RKE1 in the rest
config getter to any cluster with the `cluster-scoped-gc` lifecycle
annotation so that even if the cluster is being deleted we can still get
a kubeconfig for it to launch the cleanup job.