Fix cluster cleanup race
There is a race condition where the cluster object on the local cluster
is being deleted, but the cluster agent on the downstream cluster is not
yet cleaned up, and without this change the controller won't get a
kubeconfig for the deleted cluster and won't be able to create the
cleanup job.

To fix this for RKE1, an exception was made in ToRESTConfig to have it
return a kubeconfig even as the cluster was being deleted, as long as
the driver was RKE1. This doesn't help for the case where the driver is
"Imported" or "RKE2". It's also problematic because it means that
controllers will be unnecessarily started when the cluster is being
deleted.

This change moves the deletion check out of ToRESTConfig and into the
internal controller handling methods, and reworks the imported cluster
cleanup and node cleanup handlers to use clustermanager.ToRESTConfig and
config.NewUserContext, circumventing clustermanager.UserContext which
starts controllers. With the deletion check moved out of ToRESTConfig,
cleanup handlers can safely use it to get credentials for the cluster
and launch cleanup jobs.