Stop panic if registry CA bundle given without TLS
When provisioning an RKE2/K3S cluster with a private registry, the user
has the option to provide TLS and/or CA bundle. If the user did not
provide a TLS secret, but did provide a CA bundle, Rancher would panic.

This fixes that panic, and a user can now provide TLS secret and/or CA
bundle without issue.