Make impersonation service account more resilient
Currently, the impersonation path checks for the existence of a matching
impersonation clusterrole on the downstream cluster and then assumes
there is an impersonation service account to go with it. If for some
reason there was an interruption in between creating the clusterrole and
creating the service account, the service account won't exist and
further requests will be perpetually stuck. This change ensures that
won't happen by checking for a Not Found error when retrieving the
service account, and continuing to create the needed resources in such
an event. This also fixes an unclear error message to clarify that the
missing resource is the service account, not the secret, and to
distinguish it from another similar error message.