Ensure system and default projects get created
Currently in the hosted Rancher setup, system and default projects will
not get created for the local cluster on the tenant cluster because the
local cluster does not have a creatorID (this will be changed in
2.5.10). When the tenant cluster is upgraded (to 2.5.10 or 2.6+), then
the system and default objects will still not be created because the
`SystemProjectCreated` condition will be `true`, and `DoUntilTrue` will
not call the passed function if the condition is already true.

This change will set the relevant conditions on the local cluster to
Unknown on upgrade once so that the project controller will create
the System and Default projects if they were not created before upgrade.