Don't move namespaces to system project in hosted Rancher
Before this change, when an app was installed into a project and
namespace, the namespace was automatically moved into the project if it
wasn't already. This would cause problems in the hosted Rancher setup if
the namespace was already in the system project in the host cluster.
After this change, if the namespace is a known system-project namespace,
it is not moved to protect this case.

In addition, fleet namespaces were moved to the system project in the
local cluster automatically. In the hosted Rancher setup, this cause the
namespace to be moved back and forth between the system project on the
tenant cluster and the system project for the tenant cluster on the host
cluster. Now, if the fleet namespace for the local cluster is already in
a system project for a cluster that is not "local" then the namespace is
not moved.