Handle flapping value between rancher and fleet
When fleet bootstraps the local cluster it uses a non standard secret name
"local-cluster" because the loopback connect is handled different from
any other cluster. Due to a race condition on start it was possible that
Rancher would first create fleet-local/local cluster with no kubeconfig
secret.  If the kubeconfig secret is blank fleet would try to generate
one and would fail for the local cluster.  Instead if rancher creates
the local cluster first ensure that we also set the non standard secret
"local-cluster".