Stop cleanup agent for provisioningv2 clusters
If a provisioningv2 cluster is provisioned successfully and goes into
some type of error state, then deleting the cluster will not cleanup the
management cluster object because Rancher will try to cleanup the
cattle-cluster-agent. Rancher does not need to do this for
provisioningv2 clusters because the cluster is being deleted.

Instead, if Rancher detects that the management cluster object is
administrated by a provisioningv2 cluster, then it will not try to
cleanup the cattle-cluster-agent on delete.