Remove periods from machine name passed to machine job
For certain cloud providers, the name of the machine passed to
rancher-machine gets set to the hostname for the machine. This hostname
is then used when deploying RKE2/K3S.

If the hostname of a machine has a period in it, cloud-init will split
the hostname on '.' and set the hostname of the machine to the first
chunk. For example, if the name of the machine is
'my.really.good.machine', then cloud-init will set the hostname to 'my'.
This causes a problem in provisioningv2 if the user puts a period in the
name of a machine pool. Then all the corresponding nodes in that machine
pool will have the same name because of the chunking behavior.

This change will replace all periods in the machine name passed to
rancher-machine to dashes to avoid this problem.