Update naming of roleBindings to be unique
Problem:
The roleBindings being created might not be unique when they reference a
role like creator-cluster-owner which can apply to projects. The 2nd
user to create a project would then takeover this binding and the first
use could lose access to view the cluster

Solution:
Add a hash of the subject to the name to ensure the binding is unique
across users.