Add agent args to environment for instruction
When passing the system-agent-install instruction to the machine, none
of the agent arguments were being added to the environment. This made it
impossible to use the agent args to pass things like the PROXY
environment variables to the system-agent-installer.

After this change, the agent args are passed to the environment for the
system-agent-installer and it will pull out RKE2_ and _PROXY environment
variables.

This also includes a bump of the system-upgrade-controller that handles
environment variables for upgrades.