Stop trying to watch CRDs that have been deleted
When a user enables a node driver, there are CRDs that are are created
in Kubernetes that Rancher watches with the dynamic client. When a user
disables a node driver, the CRDs are deleted but Rancher will not stop
watching them with the dynamic client.

This change will cause the dynamic client to stop watching CRDs that are
deleted.

In addition, the machine create job would be created if the providerID
for a machine was not set or if the create job had not completed.
However, this would cause a problem if the providerID was set and the
machine create job was created.

After this change, the machine create job will be created if the
providerID has not been set.