Updated test_ingress for change from xip.io->sslip.io
Updates for 2.5.10+ and 2.6+ to use sslip.io

Reverted test_secrets, added helper for test_ingress

Reverted common.py

Removed reference from common.py

fix order of admin password

Add APIGroup to subjects

Problem:
k8s adds this field after the roleBinding is created with a default
value. This causes apply to see a difference and think the roleBinding
needs to be updated to remove the field thus causing a fight with k8s.

Solution:
Include the APIgroup when creating the subjects to stop the fightning.

Update naming of roleBindings to be unique

Problem:
The roleBindings being created might not be unique when they reference a
role like creator-cluster-owner which can apply to projects. The 2nd
user to create a project would then takeover this binding and the first
use could lose access to view the cluster

Solution:
Add a hash of the subject to the name to ensure the binding is unique
across users.

Update test for role change

switching order to fix deployment

Ensure all tags with a dash are marked as prereleases

Finalize constants, variables, and versions for v2.6.1

Update latest and stable tags in README to v2.6.1

Change consts/variables back to dev post v2.6.1

Removed reference from common.py

fix order of admin password

Add APIGroup to subjects

Problem:
k8s adds this field after the roleBinding is created with a default
value. This causes apply to see a difference and think the roleBinding
needs to be updated to remove the field thus causing a fight with k8s.

Solution:
Include the APIgroup when creating the subjects to stop the fightning.

Update naming of roleBindings to be unique

Problem:
The roleBindings being created might not be unique when they reference a
role like creator-cluster-owner which can apply to projects. The 2nd
user to create a project would then takeover this binding and the first
use could lose access to view the cluster

Solution:
Add a hash of the subject to the name to ensure the binding is unique
across users.

Update test for role change

switching order to fix deployment

Ensure all tags with a dash are marked as prereleases

Finalize constants, variables, and versions for v2.6.1

Update latest and stable tags in README to v2.6.1

Change consts/variables back to dev post v2.6.1

Finalize constants, variables, and versions for v2.6.1

Change consts/variables back to dev post v2.6.1