Clean up additional rancher namespaces
Add two additional system namespaces to the cluster cleanup agent,
cattle-fleet-system and cattle-impersonation-system.

Ensure that the 'local' namespace has the 'local' cluster as an owner
reference. This way, running `system-tools remove` on an imported
downstream cluster will cause the 'local' namespace to be automatically
removed when the 'local' cluster is removed.