Add ACE support for imported RKE2/K3S clusters
In order to support ACE for RKE2/K3S clusters, the provisioning cluster
object needed to be altered. The LocalClusterAuthEndpoint field was in
the wrong place, and this change moves it to its correct one. This will
disable ACE on RKE2/K3S clusters on upgrade, but this is acceptable
given the Tech Preview status of RKE2 provisioning.

In addition, this change will force the redeploy of the cluster-agent
and kube-api-auth when the RKE2/K3S cluster is upgraded. This is because
the daemonset of the kube-api-auth may be deleted when the downstream
cluster is upgraded. Forcing the redeploy will ensure that the
kube-api-auth containers are there when the cluster upgrade is complete.