Add kustomize for migrated apps
Helm3 apps that were installed by creating an app.project may
have had some of their resources changed with kustomize. Some
of these changes were to matchLabel selectors which are
immutable. This caused any updates to a migrated app to
result in an error because the manifests that were
being applied would try to remove the matchLabels, which
cannot be changed. Now, migrated apps will continue using
kustomize to make the same changes to the manifests as
before.