Stop cluster healthcheck on etcd restore
When and RKE2/K3S cluster is being restored from a backup, the
management cluster will go into an error state because the healthcheck
fails. While this is expected, it would be better if this did not
happen.

This doesn't happen for RKE1 provisioning because the Provisioned
condition gets set to Unknown. After this change, when and RKE2/K3S
cluster is being restored, the Provisioned condition on the management
cluster is set to Unknown to stop the healthcheck.

The cluster will go into an "Unavailable" state, but, again, this is
expected and will not produce the ugly red error messages in the UI.