Add retries to steve proxy dialer
If the cluster agent is disconnected (for instance, due to an upgrade)
and is being slow to reconnect, the transport's DialContext will fail to
get a connection to the cluster which bubbles up as an Internal Server
Error and makes it appear to the user as though Rancher itself has
crashed if they were trying to access the disconnected cluster at the
time. This change address the problem by adding retries to the dialer in
the steve proxy, which should give the agent enough time to reconnect
before a response needs to be returned to the user. This is similar to
how the dialer works in the aggregation server[1]. In this case we have
a Dialer function rather than a remotedialer Server object so we can't
use HasSession, but catching the error from the dial attempt has the
same effect.

[1] https://github.com/rancher/rancher/blob/76e743c2b38970c42e8f7a2f419c7b030c17dfb8/pkg/api/steve/aggregation/aggregation.go#L99