Copy tls-ca-additional secret to fleet-default ns
When Rancher is behind a proxy with a certificate, a tls-ca-additional
secret should be created by the user in the cattle-system namespace.
However, the machine provisioning job (for both creation and deletion)
needs this secret as well. Since this job runs in the fleet-default
namespace, the secret must also exists in this namespace.

This change will look for such a secret in the cattle-system namespace
and copy it to the fleet-default namespace on startup.