Fix disabling RKE2 feature flag
Firstly, the provisioning cluster indexer was mistakenly put behind the RKE2
feature flag. This caused an issue when deleting RKE1/imported clusters
when RKE2 feature flag was turned off.  This change puts it behind the
ProvisioningV2 feature flag.

In addition, the change adds logic to keep track of when the RKE2
feature flag can be disabled. That is, if there are RKE2 provisioned
clusters then a user should not be able to disable the RKE2 feature
flag. The lockedValue is used to the feature for easy tracking, and
the lockedValue is checked on startup and added if necessary so that
RKE2 provisioning works as expected.