Add better support for upgrading local cluster
When the local cluster is provisioned via rancherd, a hostPort is used.
When Rancher is contacted via the hostPort, the internal-cacerts setting
is used instead of the regular cacerts setting. This was causing issues
with system-agent trying to verify Rancher's CA. After this change, The
install script logic with provide the corresponding CA as necessary.

In addition, when certificates were being regenerated on Kubernetes
upgrade, the clusterIP was being used but not the hostIP. This again was
causing issues with system-agent verifying Rancher's CA because the SAN
did not include the hostIP when hostPort was enabled.

Finally, once machines were "uncordoned" the annotation was never
getting removed. This change will remove the annotation after the node
is uncordoned.