Ignore missing secrets when user's token secret is not found
Currently if a login token is missing it's corresponding secret an
error is created and the handler will be retried.  This causes a large
amount of 404 calls. This can is conservative in that it doesn't not
fix the underlying issue of why the secret is missing but only ensure
that if we hit that scenario we don't continue failing causing a large
amount of 404s.