Handle case insensitivity and roles with wildcards
When reconciling control plan roles a mapping of desired resource to verbs
is constructed and merged into the current role.  The current logic
does not consider * causing more rules to be added to the role instead
of detecting the role is already sufficient. Once the role is changed,
for the local cluster, the user controllers that since roles for role
template will delete the added rules.  This causes a flapping behavior in
writes to the same role.

Additionally the user controllers will lowercase verbs and resources names
if the roletemplate is wrong.  The control plane reconciliation was not
doing the same causing again a possible flapping behavior.