Ensure all caches are initialized before starting any controllers
During startup you may be registering handlers on controllers that have
already started.  Because of this adding the handler may be immediately
activated while other controllers are still waiting for Start() to be
called.  This causes a situation in which a handler that accesses
multiple caches could have an empty or partially populated cache.