Add support for deleting machines that failed creation
If a machine failed creation because of a user misconfiguration, then
when the user corrected their mistake a new machine would get created
but the old machine would not get removed. This change introduces
support for deleting machines without user interaction if the machine
failed creation. In addition, this change makes it possible for a user
to force deletion of a machine if the machine cannot be deleted for some
reason.

This change also cleans up the machine-status controller. A type is
created to simplify the return of a few functions.

A special note: conditions are added here to the infrastructure machine
objects. Based on some code here, it seems that they were always
intended to exist but were leftout.