Fix ingress template for `helm template` command
Without this patch, the `helm template` command generates Ingress
resources that use the `networking.k8s.io/v1beta1` APIVersion. This is
because the `helm template` command does not populate the
`Capabilities.APIVersions` set with real available resources, it only
creates a fake set of API versions that does not include resources, so
it does not have `networking.k8s.io/v1/Ingress` in the set. Since the
fake set also does not include `networking.k8s.io/v1beta1/Ingress`, we
can assume in that case that we are running in `helm template` and we
should default to the v1 version.