Allow cloud credential usage for owned clusters
If a user is an owner of a cluster, then they should be allowed to use
the cloud credential associated to that cluster in the meta proxy. This
really only applies to KEv2 and RKE2/K3S provisioned clusters because
those clusters have cloud credentials associated to them.

In addition, if a user is a owner of a cluster, they should also be able
to change the cloud credential associated to a cluster to a cloud
credential they own regardless if they have access to the old cloud
credential.