Stop passing k8s objects as interface{}
This fixes a very obscure problem that is a headache for dev local
environments. By passing the whole `object` to the formatting string,
the formatting code would try to print out the entire object. This is
totally fine in production and will not cause any problems. However, if
running Rancher in a debugger, the debugger will see a nil-pointer
dereference when printing out the DeletionTimestamp (unless the object
has been deleted). The debugger doesn't know that the panic will be
recovered and will pause execution. It will also not all execution to
continue and will terminate the go routine. This means that the error
will continue to happen.

By passing the `object` as a metav1.Object instead allows the formatting
string to jsut print out the name of the object, avoiding the issue
with the debugger.