Update project.cattle.io ingress schema
This change updates the norman ingress store to use the
networking.k8s.io/v1 Ingress API for storing Ingress objects when
possible, and modifies the project.cattle.io schema for Ingress
resources to use the k8s.io/api/networking/v1 package. This also adds
norman mappers to copy fields where needed, such as from
"defaultBackend" to "backend", so that no matter which storage API is
used, the right data exists in the object. See the deprecation guide[1]
for notes on how the fields are translated.

[1] https://kubernetes.io/docs/reference/using-api/deprecation-guide/#ingress-v122