Add the ability to defer the start of user controllers
This PR lays the framework to allow user controllers to watch for
conditions before they start their caches/controllers. This is to allow
performance gains if the user is not using any legacy features.

Also included is the ability to limit which namespaces a controller is
watching which will be used by the impersonation logic in a follow on
PR.