Defer start of approuter user controller
Only start the approuter user controller if configured to a different API
endpoint.  This feature does not seem to be used anymore as the default
setting points to a service that doesn't exist (https://api.lb.rancher.cloud/v1).
If the setting points to the default we know it doesn't work so only
start the user controller if it's set to some other non-empty string.