Add compatibility helpers for ingress controllers
This change adds a package called ingresswrapper which provides
utilities that help make controllers able to use either the
extensions/v1beta1 Ingress API or the networking.k8s.io/v1 Ingress API,
depending on the version of the kubernetes API server the controllers
are running on. Kubernetes 1.18 only supports the extensions/v1beta1
Ingress API, while Kubernetes 1.22 only supports the
networking.k8s.io/v1 Ingress API. The compatibility layer defaults to
using the networking.k8s.io/v1 API when it is available.

The wrapper package provides two main helper objects that the
controllers can use. One is the Ingress interface, which is just a
controller-runtime client Object, which itself is a meta interface for
the apimachinery runtime Object and the meta/v1 Object. It allows direct
access to the object's TypeMeta and ObjectMeta data and can be used when
information from the object's Spec field is not needed. The other is the
CompatIngress struct, which inlines the networking.k8s.io/v1/Ingress
struct and provides methods to convert any Ingress object. This is
useful when the actual Spec of an Ingress object is needed. The data
structure is the same as for a networking.k8s.io/v1/Ingress object, so
when support for Kubernetes 1.18 is eventually removed, we should be
able to simply replace `CompatIngress` with `knetworkingv1.Ingress` in
every controller.

The wrapper also provides interfaces and wrapper methods to help with
setting up controllers.