Better sync of infra machine conditions
The CAPI machine controller will automatically sync the Ready condition
on the infra machine object to the InfrastructureReady condition on the
machine. This was causing issues when setting the InfrastructureReady
condition on the Machine (which we needed to to so the CAPI machine
controller would pick up the changes).

Now, the Ready condition on the infra machine and the
InfrastructureReady condition on the machine are set to the same thing
to avoid this confusion.