Stop overwriting ACE config for imported RKE2/K3S clusters
When an imported cluster is created, it is created as a provisioningv2
cluster. However, after an imported cluster is recognized as an RKE2/K3S
cluster, then the management cluster is edited. This caused a problem
because the provisioning cluster controllers would overwrite the ACE
config on the management cluster as soon as the management cluster was
edited.

Now, the ACE config is deleted before the provisioning controller saves
the management cluster preventing the ACE config on the management
cluster from being overwritten.