Always start the Secrets controller
The impersonation mechanism relies on the Secrets controller being
started on the user context when a proxy request is made. Before
6b2e1359, the controller was incidentally being started as part of
registering other controllers. With 6b2e1359, the certsexpiration
controller was the last to always start it on startup and now only
conditionally starts it for RKE clusters. This change adds back the
Secrets controller startup to the cluster context unconditionally so
that it is ready the first time impersonation.GetToken is called.