Skip cloud cred check if it is not changed
A cluster owner should be able to edit a cluster without changing the
cloud credential. Therefore, if the edit isn't changing the cloud
credential, then Rancher shouldn't validate that the user has access to
the cloud credential.

After this change, Rancher validates that the user has access to the
cloud credential only if they are trying to update it.