ISSUE 35694: Support ingressClassName in Helm Vals
Added a Jinja2 tempalte that injects ingressClassName into the ingress.yaml spec.
This remedies an issue where the created Ingress returns 404's when no default
IngressClass is found by the Ingress resource. This also enables support for
users choosing which of multiple possible Ingress classes to use when installing
with Helm.