Add currentVersion query parameter for templates
A currentVersion query parameter can now be passed when making
templates requests to norman. This parameter will cause any
version matching it to be included in versionLinks. This is
being done so versions that are not compatible with current
setup but are already in use do not have to be filtered out.
This is required to facilitate the UI, otherwise the version
is treated as though it doesn't exist, despite being in use.