Check if matching dashboard app exists
Prior, if a dashboard app already existed with the same ID in a
namespace that matched the targetNamespace of the
apps.project.cattle.io being created, they would be linked to
the same helm resources. This means changing one changed the
other. Now, an apps.project.cattle.io cannot be created if
a apps.catalog.cattle.io already exists with the same ID in
a namespace matching its targetNamespace.