Fix EKS nodegroup deletion on AWS API error
In the current implementation, an error was not checked, which led to all
nodegroups being deleted when an error was received from the AWS API when
listing nodegroups. Now, the error is checked and no deletion should occur.

In addition, to avoid setting the upstreamSpec to nil on refresh, each config on
 the `upstreamConfig` object is nil-checked. At least one of them should not be
 nil for the refresh to proceed.