Check for nil clusterSpec in kev2 refresher
Previously, when an error was received in the upstream refresher, only the
`upstreamConfig` was nil-checked. However, if retrieving the upstreamSpec
failed, then `upstreamConfig` would not be nil, but the corresponding cluster
config would be. Therefore, in addition to checking the `upstreamConfig` for
nilness, the individual cluster configs are also checked. At least one of them
should be non-nil for syncing to proceed.