Restart agent pods after upgrade to 1.22
When a cluster is created with version 1.18 or lower, various
controllers will start a watch for the extensions/v1beta1/Ingress type.
Without this patch, if that cluster is then upgraded to 1.22, those
watches will start to fail because that API is no longer available, and
the controllers using those watches will stop syncing because they don't
see events for the new Ingress type, networking.k8s.io/v1/Ingress.

This patch adds a check to the clusterstats controller, where the
cluster version status is updated, and annotates the cluster agent
deployment template which forces a restart of the cluster agents. Simply
forcing the agent to be reapplied via the clusterdeploy controller does
not necessarily fix the problem because if there is no change to the
cluster agent deployment template then the pods won't actually be
restarted.

The version checks for an original version of 1.21, rather than 1.18.
This is because it's possible that a cluster was created with version
1.18 or below, then upgraded to version 1.21 or below, where the old
extensions/v1beta1 API would still be available and the controllers
would still work. But then the final upgrade to 1.22 would break the
controllers. So, assume that any upgrade to 1.22 needs the agents
restarted.