Don't block controller thread with deferred start
If two controller goroutines both run deferred start it is possible to deadlock
as one goroutine is holding the start lock and waiting for the controllers
to start and the other gorouting is trying to get the start lock to wait
for the controllers to start.  Since both can not move forward the
controller start is deadlocked.