Adding CSI Proxy config to Windows RKE2 Install script.
Windows requires a CSI Proxy windows service to be executing on the host for enabling CSI drivers to allocate resources on Windows nodes. We have already added that capability to Wins to support installing and starting the service. The configuration for that service needs to be driven by Rancher. We are currently defaulting this to enabled by default. If we later determine to make this configurable by Rancher, then we just need to update the this code in Rancher and the corresponding unit test to test that behavior.

Signed-off-by: Jamie Phillips <jamie.phillips@suse.com>