Ignore cluster agent restart for local cluster
There is no cluster agent on the local cluster so it does not need to be
restarted. Attempting to look up the local cluster agent and update it
was failing because the Deployment object returned from Get was non-nil
but empty, so trying to Update that object was invalid. This change 1)
skips the agent restart for local clusters, 2) moves the restart to
after the cluster update so that any errors that occur don't interfere
with saving the version on the cluster object, and 3) fixes the
deployment lookup to check for Not Found rather than nil.