Accept preexisting incompatible app versions
Prior, if an app that was installed prior was rendered incompatible
due to a rancher upgrade, updates to the app would fail. This would
happen even if the update was not to the app version. Now, apps can
be updated if their current version is incompatible as long as they
are not updating to another incompatible app version.