Add SUC chart version environment variable
The SUC chart version was buried in the Rancher code. This change make the
version a setting and adds an environment variable that is similar to the other
versions that are needed.

The goal for this change is to make updating this versio similar to how other
versions are updated, so it is less likely to be missed.