Leave out bootstrap secret on delete
When deleting a machine, if the bootstrap secret is optional and not found,
then the name of the secret will be changed when passing it to the job. This
causes Wrangler Apply to delete the job and recreate it with the new name, which
in turn may cause the job to hang because the process of create -> delete ->
create is too fast for Kubernetes to handle.

Now, the volume and volume mount are only include if the bootstrap secret is
required. This keeps the job from being deleted and recreated when the bootstrap
secret cannot be found.